Fix duplication of metadata in header/footer"
The metadata present in the footer of a CaRTed file would be extracted into the cartmeta file. This in turn would be written to the cart header if/when reCaRTed, which would cause duplication of information, as well as a chance that the data might be stale following a change.